moving change notification settings intent handler to separate file
